Fixed Round, Floor and Ceiling returning a Long rather than a Double. Closes #131

+++ src/main/java/vazkii/psi/common/spell/operator/number/PieceOperatorCeiling.java
@@ -34,7 +34,7 @@ public class PieceOperatorCeiling extends PieceOperator {
 	public Object execute(SpellContext context) throws SpellRuntimeException {
 		Double d = this.<Double>getParamValue(context, num);
 
-		return Math.ceil(d);
+		return Double.valueOf(Math.ceil(d));
 	}
 
 	@Override
